+++ code/src/sixsq/nuvla/server/resources/user.clj
@@ -140,9 +140,7 @@ requires a template. All the SCRUD actions follow the standard CIMI patterns.
 ;; requires a user-template to create new User
 (defmethod crud/add resource-type
   [{:keys [body form-params headers] :as request}]
-
   (try
-
     (let [authn-info   (auth/current-authentication request)
           body         (if (u/is-form? headers) (u/convert-form :template form-params) body)
           redirect-url (get-in body [:template :redirect-url])
@@ -228,6 +226,7 @@ requires a template. All the SCRUD actions follow the standard CIMI patterns.
       (-> merged
           (dissoc :href)
           (u/update-timestamps)
+          (u/set-updated-by request)
           (crud/validate)
           (db/edit request)))
     (catch Exception e